Revert "Merge branch 'main' into grpc_configurable"
This reverts commit 673119c69c3afc1ed5610fd3f5805db6e3b63651, reversing
changes made to 3ed780c09b26c137f04e1270a2a86763970e5a8c.

+++ mobile/library/swift/Stream.swift
@@ -37,7 +37,7 @@ public class Stream: NSObject {
   ///                        callback.
   ///
   /// - returns: This stream, for chaining syntax.
-  public func readData(_ byteCount: Int) -> Stream {
+  public func readData(_ byteCount: size_t) -> Stream {
     self.underlyingStream.readData(byteCount)
     return self
   }
